requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
scipy>=1.11.0
pyyaml>=6.0

# SAP integration
pyrfc>=3.0.0  # Optional: for direct RFC connection

# Visualization & Dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.28.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database (for logging)
sqlalchemy>=2.0.0
pymongo>=4.5.0  # Optional: for MongoDB logging

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Monitoring & Logging
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Cloud deployment
boto3>=1.28.0  # AWS SDK
google-cloud-storage>=2.10.0  # Optional: for GCP
azure-storage-blob>=12.17.0  # Optional: for Azure

# Additional ML tools
optuna>=3.3.0  # Hyperparameter optimization
shap>=0.42.0  # Model interpretability
